# more for loop = more speed