Added more loops which allows better caching behavior, sped up code by about 12% (73->82 GFLOPs) on my machine (4 cores) using 
```gcc -O2 gemm.c -ffast-math -lpthread -march=native```
and about 70% faster (130 -> 220 GFLOPs) when using -O3 optimization compared to previous implementation

Here's the output files testing the code
[before.txt](https://github.com/geohot/tinygrad/files/9333971/before.txt)
[after.txt](https://github.com/geohot/tinygrad/files/9333970/after.txt)
[before_o3.txt](https://github.com/geohot/tinygrad/files/9333972/before_o3.txt)
[after_o3.txt](https://github.com/geohot/tinygrad/files/9333969/after_o3.txt)