@geohot 2022-09-11
This isn't seriously maintained, it's an example to get max perf on Ryzen 5950